simplify scroll to bottom logic for logs
