+++ frontend/src/components/tasks/TaskDetails/LogsTab/Conversation.tsx
@@ -1,4 +1,4 @@
-import { NormalizedConversationViewer } from '@/components/tasks/TaskDetails/NormalizedConversationViewer.tsx';
+import { NormalizedConversationViewer } from '@/components/tasks/TaskDetails/LogsTab/NormalizedConversationViewer.tsx';
 import {
   useCallback,
   useContext,
@@ -10,19 +10,17 @@ import {
 import { TaskAttemptDataContext } from '@/components/context/taskDetailsContext.ts';
 import { Loader } from '@/components/ui/loader.tsx';
 
-type Props = {
-  conversationUpdateTrigger: number;
-  handleConversationUpdate: () => void;
-};
-
-function Conversation({
-  conversationUpdateTrigger,
-  handleConversationUpdate,
-}: Props) {
+function Conversation() {
   const { attemptData } = useContext(TaskAttemptDataContext);
   const [shouldAutoScrollLogs, setShouldAutoScrollLogs] = useState(true);
+  const [conversationUpdateTrigger, setConversationUpdateTrigger] = useState(0);
 
   const scrollContainerRef = useRef<HTMLDivElement>(null);
+
+  // Callback to trigger auto-scroll when conversation updates
+  const handleConversationUpdate = useCallback(() => {
+    setConversationUpdateTrigger((prev) => prev + 1);
+  }, []);
 
   useEffect(() => {
     if (shouldAutoScrollLogs && scrollContainerRef.current) {
