+++ frontend/src/components/tasks/TaskDetails/LogsTab/NormalizedConversationViewer.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { Bot, Hammer, ToggleLeft, ToggleRight } from 'lucide-react';
-import { Loader } from '@/components/ui/loader';
+import { Loader } from '@/components/ui/loader.tsx';
 import { executionProcessesApi, withErrorHandling } from '@/lib/api.ts';
 import { MarkdownRenderer } from '@/components/ui/markdown-renderer.tsx';
 import type {
